backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
python-decouple
celery
redis
razorpay
stripe
reportlab
openpyxl
pandas
numpy
python-dateutil
pytz
requests
gunicorn
whitenoise
django-extensions
django-debug-toolbar
drf-yasg
djangorestframework-simplejwt
django-jazzmin